+++ node/src/main.ts
@@ -1,5 +1,5 @@
 export type { RTCDataChannelLike, RTCPeerLike, ChannelReliability } from './peer/RTCPeer'
-export { createPeerConnection, NodeDataChannelPeer, NodeDCDataChannel } from './peer/RTCPeer'
+export { createPeerConnection, NodeDataChannelPeer, NodeDCDataChannel, WeriftPeer, WeriftDataChannel } from './peer/RTCPeer'
 
 export type { RTCAdapters, PeerNegotiatorCallbacks, HostNegotiatorCallbacks } from './peer/NegotiationCore'
 export { PeerNegotiator, HostNegotiator } from './peer/NegotiationCore'